Move backup buttons to Settings > System
Add icon to backup card
